+++ cypress/support/step_definitions/annual-review/scheduler/an_annual_review_date_has_been_set.js
@@ -1,4 +1,4 @@
-And("an annual review date has been set", async () => {
+Then("an annual review date has been set", async () => {
   cy.task("db", {
     operation: "list.findFirst",
     variables: {
