+++ src/server/models/list.ts
@@ -3,7 +3,7 @@ import { logger } from "server/services/logger";
 import { isGovUKEmailAddress } from "server/utils/validation";
 import { prisma } from "./db/prisma-client";
 
-import { CountryName, CurrentAnnualReview, List, ListCreateInput, ListUpdateInput, ServiceType } from "./types";
+import type { CountryName, CurrentAnnualReview, List, ListCreateInput, ListUpdateInput, ServiceType } from "./types";
 import { subMonths } from "date-fns";
 
 export async function findListById(listId: string | number): Promise<List | undefined> {
@@ -86,25 +86,6 @@ export async function findListByAnnualReviewDate(annualReviewStartDate: Date): P
   }
 }
 
-export async function findListsWithoutNextAnnualReview() {
-  try {
-    const result = (await prisma.list.findMany({
-      where: {
-        nextAnnualReviewStartDate: null,
-      },
-      include: {
-        country: true,
-      },
-    })) as List[];
-
-    logger.debug(`${result.length} lists without nextAnnualReview found`);
-    return result;
-  } catch (error) {
-    logger.error(`findListsWithoutNextAnnualReview Error: ${(error as Error).message}`);
-    return { error: new Error("Unable to get lists in annual review") };
-  }
-}
-
 export async function findListsWithCurrentAnnualReview(): Promise<Result<List[]>> {
   try {
     const result = (await prisma.list.findMany({
@@ -126,20 +107,6 @@ export async function findListsWithCurrentAnnualReview(): Promise<Result<List[]>
     logger.error(`findListsInAnnualReview Error: ${(error as Error).message}`);
     return { error: new Error("Unable to get lists in annual review") };
   }
-}
-
-export async function findFirstPublishedDateForList(listId: number) {
-  return await prisma.event.findFirst({
-    where: {
-      listItem: {
-        listId,
-      },
-      type: "PUBLISHED",
-    },
-    orderBy: {
-      time: "asc",
-    },
-  });
 }
 
 export async function createList(listData: {
@@ -256,20 +223,4 @@ export async function updateListForAnnualReview(
     logger.error(errorMessage);
     return { error: new Error(errorMessage) };
   }
-}
-
-export interface ListWithFirstPublishedDate {
-  listId: number;
-  oneYearAfterFirstPublishedDate: Date;
-}
-
-export async function addAnnualReviewToList({ listId, oneYearAfterFirstPublishedDate }: ListWithFirstPublishedDate) {
-  return await prisma.list.update({
-    where: {
-      id: listId,
-    },
-    data: {
-      nextAnnualReviewStartDate: oneYearAfterFirstPublishedDate,
-    },
-  });
 }
